(PC-7698) Remove `interact` script
We can now use `flask shell` instead.
